+++ protected/humhub/modules/queue/migrations/m190211_133045_channel_length.php
@@ -1,0 +1,47 @@
+<?php
+/**
+ * @link https://www.humhub.org/
+ * @copyright Copyright (c) 2019 HumHub GmbH & Co. KG
+ * @license https://www.humhub.com/licences
+ */
+
+use yii\db\Migration;
+
+/**
+ * Class m190211_133045_channel_length
+ */
+class m190211_133045_channel_length extends Migration
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function safeUp()
+    {
+        $this->alterColumn('queue', 'channel', $this->string(50)->notNull());
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function safeDown()
+    {
+        echo "m190211_133045_channel_length cannot be reverted.\n";
+
+        return false;
+    }
+
+    /*
+    // Use up()/down() to run migration code without a transaction.
+    public function up()
+    {
+
+    }
+
+    public function down()
+    {
+        echo "m190211_133045_channel_length cannot be reverted.\n";
+
+        return false;
+    }
+    */
+}